Added converter tool for docker images
This commit provides a tool which can download a docker image
form the dockerhub or another docker registry using curl.
This image is converted to a trustx GuestOS image by converting
the docker manifest json to a protobuf Guestos config and
squashing the docker layers to a squashfs image using mksquashfs.