Make features ipc_ns, input key, radio routing table hardware dependend
Depending on the kernel config of the device some features are
diffrent or not available at all. Thus, we moved this to the
device specific hardware hw_*.c implementation.
Also we implement a routing table flush, which is used in a0
to clean up the routing tables and set a generic routing rule.