tpm2d: Implemented secureboot policy (PCR7) for fde key read auth
This patch introduces the a policy for the nvmcrypt submodule,
which depends on the value of PCR7. This policy is used as
read auth for the corresponding nvindex in case the submodule
ois initialized with the use_secure_boot_policy paramater set to
true.

Further some minor fixes are included and also for debugging
the undfine space command is implemented which allows a dmcrypt_reset
throug the control interface.