cmld: use cgroup namespace if available
This patch implments the cgroup namepsace. Therfore it also mounts
the subsystems to separte folders. There for each container
a new subfolder is created which the host (cmld) controls,
e.g., /sys/fs/cgroup/devices/00000000-0000-0000-0000-000000000000.
In that subfolder another child subfolder is created which is
in control of the container, e.g.,
/sys/fs/cgroup/devices/00000000-0000-0000-0000-000000000000/child.
In the child subfolder the container is than namespaced, without
the ability to acces the upper folder which the cmld uses to enforce
resource limits.