Implemented IMA-style measurement list (ML) for containers
This patch provides the implementation to provide an IMA-style
measurement list (ML) for container images. The cmld (guestos.c)
measures (generates the hash) of each read-only image before starting
the container and sends it to the tpm2d through the tss.h interface.
The tpm2d maintains the measurement list and extends new measurements
from the cmld to the TPM's PCR11, thus not interfering with the
IMA default in PCR10.

The ML is send at the end of the attestation request in the tpm2_control
interface to a remote verifier.

A sample ml_entry looks like this:

ml_entry: "11 4bf6d86968043a73640ebeb82d6a67481bc66f2ae7a9f5589d57fdebf3011b20 \
 ima-ng sha256:ace88b9a1573d8e0d2555d89b408b8a15d36f4d2ed1468d46143e4e0e9ff9579 \
 /data/cml/operatingsystems/trustx-coreos-1/root.img"