c_cgroups: ensure the removal of container cgroups on cleanup
If the container is killed and previously some cgroups where created,
this could lead to left overs which caused the actual cleanup code
to fail. E.g. systemd is creatings sub cgroups which caused
an EBUSY in the cleanup code. This patch removes all cgroups of a container
recursively to avoid the EBUSY on removel of the parent cgroup.