daemon: avoid Assertion `if_name' failed by sanitizing mac
If the user sets invalid mac address kernel will reject the
creation of the veth pair with "99: Address not available".
In that case, an assertion for the if_name of that interface
is triggered in c_net.c causing a fatal abort.

This patch gives only sanitized mac addresses out of the
config sub module, overwriting the user provided misconfiguration.