Add support for coding style formatting
This patch adds support for clang's source code formatter.
With this tool a developer can automatically fix the code style
of a source file so that the code fits the project's guidelines (e.g.
tabs over spaces, column width, bracket positioning, etc).

The chosen style is a modified version of the Linux kernel coding style
(https://www.kernel.org/doc/html/v4.10/process/coding-style.html).
The modifications are as follows:
- the return type of a function's definition is put on a separate line
- line width: 120 instead of 80 columns
- disabled support for custom foreach macros (since our code base
  doesn't have any)

To automatically format the source code, the clang-format needs to
be installed on the host. Then it can be started with the `format-code.sh` script.

Alternatively, the tool can be executed automatically by the IDE/text
editor when a file is saved (or while typing):
- CLion has in-built support
- Eclipse has a plugin
https://marketplace.eclipse.org/content/cppstyle
- a vim plugin https://github.com/Chiel92/vim-autoformat

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>