Protect the SOCK_PATH macro from clang-format
This patch inserts comments (// clang-format off and // clang-format on)
that guard all prevents clang-format to format usages of the SOCK_PATH macro.

The SOCK_PATH is not allowed to be formatted by clang-format.
If formatted, the macro's semantics are broken and trustme cannot boot.
Further damages are not ruled out.

The main reason is that the macro SOCK_PATH is used to produce and
concatenate strings. Therefore it is white-space sensitive.
However, clang-format does not know this and formats SOCK_PATH as if it
were a function call.

For example:
- SOCK_PATH(scd-control) will normally be expanded to
  CMLD_SOCKET_DIR"/cml-""scd-control"
- but clang-format formats it to SOCK_PATH(scd - control) thinking it is
  a function call with arithemtic expression. Then the macro
  expands to CMLD_SOCKET_DIR"/cml-""scd - control" which is a different
  string and breaks the underlying semantics.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>