Revert protobuf and $SRC_FILES makefile dependency order
Commit 4b08b62 erronously removed the `$(SRC_FILES): protobuf`
dependency in the Makefile. Instead the commit substituted it
with `artifact: protobuf $(SRC_FILES)` for all artifacts that
depend on $(SRC_FILES).

This caused a race condition when
make was executed with multiple concurrent jobs. Make would try
to build the c_service.pb-c.c source which was not yet
generated by the `probobuf` task.
As a result the build would fail.

Example output of the error on the CI server:

| NOTE: make -j 96 -C service service-static
| make: Entering directory 'REDACTED'
| protoc-c --c_out=. c_service.proto
| make: *** No rule to make target 'c_service.pb-c.c', needed by
'service-static'.  Stop.
| make: *** Waiting for unfinished jobs....
| protoc-c --c_out=. container.proto
| protoc-c --c_out=. control.proto
| protoc-c --c_out=. guestos.proto
| make -C common protobuf

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>