Fix a heap buffer overflow
The problem is with the following statement:

msg.exec_args = mem_alloc(sizeof(char *) * argc);

The expression sizeof(char *) * argc could overflow.
Furthermore argc is controlled by an outside attacker
who can execute the control binary.
As a result less space for the msg.exec_args buffer could be allocated.
Then in the following while loop, the buffer could be overflown with
arbitrary content. This bug is exploitable.

The fix is to check for an integer overflow in the multiplication
before allocating the memory. This can be done with the
__builtin_mul_overflow(a, b, c) bultin-functions which is
part of gcc 5+ [1] and clang 3.4+ [2].

References:
[1]: https://gcc.gnu.org/gcc-5/changes.html#c-family
[2]: http://clang.llvm.org/docs/LanguageExtensions.html#checked-arithmetic-builtins

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>