cmld: activate user namespace for containers
This patch activates the user namespace according to the container
configuration file (default = on). A new container submodule c_user
is introduced which handles the uid and gid mappings in the root
namespace. Each container is mapped to a new range of 10000 uids
starting from uid 30000.
The c_user module also provides an interface which is used by c_vol
to shift uids and gids of directories. This feature depends on the
non-mainline shifts.