Add a dedicated folder for utility scripts
This commit adds a dedicated folder for utility scripts that are using
during the development/testing workflow (e.g. for formatting the source
code, running tests, etc.). It is cleaner to have all such scripts in a
separate place that does not pollute the repository's root.

Furthermore, the format-code.sh scripts is now moved inside
scripts/format-code.sh; Therefore you need to adjust the settings of
your tools if they relied on the format-code.sh script.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>