Fix unhandled call failures and NULL ptr dereferences
If gettimeofday fails, the &tv ptr is still dereferenced when
localtime(&tv.tv_sec) is called. `struct timeval tv` is declared on the
stack. This means localtime(&tv.tv_sec) might lead to undefined
behaviour.

Similarly, tm is declared as a pointer on the stack, which is
initialised by tm = localtime(&tv.tv_sec). If localtime fails, tm is set
to NULL. However, strftime dereferences tm even if it's NULL.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>