Fix a leaking file descriptor
If setting the cipher for CSR private key encryption fails, the control
flow is transferred to the `error goto label`, which frees the memory
allocated objects, but does not close the fp file descriptor for the
private key file.

Therefore the private key's file descriptor leaks.

This commit fixes the issue.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>