Fix a memory and file descriptor leak
The nl_sock_t *ret object is allocated on the heap. Also the ret->fd is
an opened socket file descriptor.

However, if we fail to configure the socket for different  protocol
versions (UEVENT, ROUTE, XFRM) we return from the function without
freeing the resources. Hence both a memory leak and a file descriptor
leak occur.

This commit fixes the issue.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>