Fix memory leaks in container_new
Everytime a new container is started the config_filename and images_dir
string buffers leak. They must be freed because after creation they are
no longer used. During the lifetime of the container, the contents of
both buffers are still available. They are duplicated inside
container_new_internal().

This commit also fixes 3 other memory leaks that occur when we
prematurely return from container_new because of a failer. To be more
specific:
- the container_config_t *conf object must be freed if we cannot find a
  GuestOS for the container
- the config_filename and images_dir buffers need to be freed if we have
  a version mismatch for the container's GuestOS

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>