Fix a memory leak in daemon's GuestOS code
When purging a GuestOS object we allocate a mount_t *mnt entry to get the
GuestOS's image URLs and delete them. However, after we are done purging
the GuestOS object we return without freeing the allocated mount_t *mnt
entry. This causes the daemon to leak memory everytime we request a
delete of an GuestOS.

A similar bug occurs in the iterate_images_start helper function. We
allocate a mount_t *mnt entry and check if we can iterate over the images. If
we can't we return without freeing mount_t *mnt and cause a leak.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>