Use a default value for container's vnet MAC if generation fails
It the read from /dev/urandom fails for some reason, we will fail to
initialise the uint8_t mac[6] array. Since the array is declared on the
stack and not initilised, the next call to mem_printf will cause
undefined behaviour. This could be used for crafting a write primitive.

This commit intialises the uint8_t mac[6] array with
{0xff, 0xff, 0xff, 0xff, 0xff, 0x00} to get rid of the undefined
behaviour and adds a log message to signal that we failed reading from
/dev/urandom.

There is no motivation behind the chosen default value for mac[6]
other than the last byte was intentionally picked to be a terminating 0x00.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>