daemon/smardcard <-> scd: Introduce verify using buffers for data, sig, cert
This patch introduces a new scd command code for verifying data.
Analogues to the already existing command using files this, we
now can verify buffers and send data, sig and cert through the socket,
serialized through protobuf. This is the first step to allow scd
to run containerized outside of the CML.