Harden mem_free to prevent use-after-free and double-frees
This commit:
1) changes the mem_free wrapper of free to a macro which
   protects against NULL ptr corruption, use-after-free, double-free
   and similar exploitable problems;
2) fixes a use-after-free vulnerability in c_cgroups.c;
3) fixes a null-format-string bug in converter.c.

The reason why we change the mem_free function to a macro and not extend
the function itself is backwards compatibility. Setting a freed pointer
to NULL requires passing a pointer to a pointer as function argument.
Specficially:

void mem_free(void *ptr) {
    free(ptr);
    ptr = NULL; // Does not work, broken
}

needs to be extended to

void mem_free(void **ptr) { // No longer backwards compatible
    free(*ptr);
    *ptr = NULL;
}

Therefore we change to:

 #define mem_free(ptr) \    // Backwards compatible and sets the ptr to NULL!
    free(ptr); ptr = NULL

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>