Create a skeleton for a unit testing harness
This commit initialises a bare skeleton for a testing harness.
Currently the harness is enabled only for the common/ subproject. Inside
the common/ directory, the developer can execute `make test` and run the
tests (currently only one failing test is implemented, more to come
soon).

The testing harness is based on the munit C unit testing framework [1].
The reasons for choosing this framework are as follows:
* simple API without an ugly DSL based on macros;
* has only 2 files - munit.c and munit.h;
* is agnostic to the build system and used compiler;
* supports grouping of test cases into suites;
* permissive license;
* simple enough to extend manually (that's the reason I have included
  munit.c and munit.h directly and not as a git submodule; I plan to
  extend them quite significantly, i.e. they will become a fork and
  diverge from the upstream version)

It is planned that the harness will be enabled also for the rest of the
subprojects.

References:
[1] https://nemequ.github.io/munit/

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>