Extend munit to assert for received signals
The commit extends the munit testing framework to check if a test
triggers a certain signal.

This is useful when we want to test certain protections (e.g. if a
protection macro triggers SIGABRT when it detects integer overflow).

I've added support only for SIGINT, SIGABRT, SIGKILL, SIGSEGV and SIGALRM.
However, it is trivial to add new signals. Just look at the patch diff
to understand how.

To use this new feature inside munit, add
MUNIT_TEST_OPTION_RECV{SIGNAME} to the test case's options. For example,
to indicate that a test MUST trigger SIGALRM, we specify:

static MunitResult
test_func(...)
{
    // some operations
    // something here triggers SIGALRM
}

struct MuintTest tests[] = {
    {
        "/the name of the test",
        test_func,
        setup_func,
        tear_down_func,
        MUNIT_TEST_OPTION_NONE | MUNIT_TEST_OPTION_RECV_SIGALRM,
        NULL
    },
    ....
}

The test will now fail if it doesn't trigger SIGALRM.
The test will succeed only and only if it triggers SIGALRM.

Important: munit terminates failing assertions with SIGABRT. So, if we
want to assert that a given function triggers SIGABRT, we shouldn't be
asserting anything else inside that test case (because if the assertion
fails, the wrong SIGABRT might be swallowed).

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>