Protect struct allocation wrappers against overflow
The problem
---

This commit protects the:
- mem_new(struct_type, n_structs)
- mem_new0(struct_type, n_structs)
- mem_renew(struct_type, ptr, n_structs)
convience wrappers for allocating struct objects against
integer overflow inside the sizeof(struct_type) * n_structs calculation.

There are places in the codebase where n_structs is calculated from
parameters that can be controlled from the outside (i.e. by a potential
attacker). Hence, it is possible to cause the
sizeof(struct_type) * n_structs expression to overflow and allocate less
memory than we had in mind. As a result instructions that follow will
access unallocated heap regions, leading to buffer overflow problems.

The solution
---
I've extended the wrapper macros to check for an overflow before
allocating the desired amount (using the __builtin helpers). If
overflow is detected, the program is aborted.

Because the macros need to return a value and I didn't want
to re-declare them as functions, I have taken advantage of
gcc's statement expressions [1], which clang supports also by default [2].
I have used the __extension__ intrinsic to surpress warnings that this
is not an ANSI-C feature.

Additionally, I have added unit tests to make sure that the overflow
protection is triggered. As a bonus, I have also added unit tests for
the rest of the functions from mem.h

References:
[1]: https://gcc.gnu.org/onlinedocs/gcc/Statement-Exprs.html
[2]: https://clang.llvm.org/doxygen/classclang_1_1StmtExpr.html

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>