Fix type coercion and cast errors
This commit fixes bad type coercions and casts caught by the static
analyser. The fixes preserve the original control flow.

For example, the tss_hash_algo_get_len_proto function's return type is
HashAlgLen which the compiler resolves to unsigned int. However, the
function returns -1 on error and that is being coerced to
MAX_UNSIGNED_INT.

Another example is the file_read function which returns -1 on error and
a positive int on success. However, many call-sites declare
size_t len = file_read(...) and loose the ability to check for errors.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>