daemon/download.c: handle local "file://" urls
This allows to use download mechanism for guestos from local
source. This could be used to specify an update url inside
of a container shared in cmls mountns. Thus, e.g., c0 could
be used to download an provide updates for CML.
This is the first step to get rid of external binaries used
form busybox inside of CML, in this case wget.