Fix NULL ptr dereferences in c_cgroups.c
NULL ptr dereference occurs when we allow all cgroups inside a container.
I tested and validated this bug with address sanitizer.
This same bug is also triggered when we deny all cgroups for a
container (instead of allowing all cgroups).

The function c_cgroups_devices_allow_all calls c_groups_dev_from_rule
with the rule "a". Thus char *dev will equal to NULL which is dereferenced
by char *maj_str = strtok_r(...).

The same situation can happen if min_str is NULL and is dereferenced by
strncmp(..).

This commit fixes the issue by returning (-1, -1) if the passed rule
string does not match the regex: ".+ +.+:+.+.*". (quotes for clarity,
not part of the regex). I implemented this by checking that every call
to strtok_r succeeds. If it fails, we abort the parsing and return -1.

Examples:
"c 10:232 rwm" returns (10, 232)
"b 7:* rwm" returns (7, -1)
"a" returns (-1, -1)
"aa a a" returns (-1, -1)

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>