Add CI support for clang-format
We want to have the CI check if the code is formatted or not. If a patch
violates the coding style, the pipeline should fail before building a
whole image and reject the patch.

We use clang-format to enforce code styling across the project. The
problem with clang-format is that its output differs based on the
version that is used (clang-format's version are generally not
backward-compatible). Therefore if developers use different versions,
their and the CI's output might mismatch.

To resolve this problem and have consistent code formatting both locally
and on the CI server, this commit adds a precompiled x86_64 Linux clang-format binary.
This binary is used by the `scripts/format-code.sh` script which can be
used by developers to format the code before committing.

In addition, this commits adds
`scripts/ci/check-if-code-is-formatted.sh`.
The CI uses this script to check if the code is formatted or not.

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>