Enable a new multi-staged CI pipeline
This commit adds a **preliminary** support for a new multi-staged CI pipeline.

This new design is not fully ready:
- some of the new stages work (e.g. until the yocto stage, inclusive,
  everything should work properly)
- but others haven't been implemented yet (e.g. the integration tests,
  documentation generation, etc.); I have inserted TODOs to signal such
  places.

One drawback of the current implementation is that it uses a Dockerfile
from the scripts/ci directory. This Dockerfile should be in the
trustme-build repository instead. I haven't had enough time to do this
(since it's easier to test by having the Dockerifle in this repository)

Signed-off-by: Vasil Sarafov <vasil.sarafov@aisec.fraunhofer.de>