scd: cache the usbtoken auth code even when it gets locked
We need to be able to forward the unlocked usbtoken to a container so it
can be used from itside it. To be able to re-unlock the token if it gets
locked we must enable the software inside the container to request the
scd to do so. That way, the token pin must not be re-entered and it does
not get exposed to the container software/user.