control/control: fixed control flow for container specific commands
This path checks if the container with the given name or uuid
really exists before sending the actual command to cmld.
This way also the change_pin command would not ask for passwords
if the container does not exist.

This fixes GitHub Issue #148: "control: change_pin does not check
container-uuid before asking for password"