daemon: properly mount container subtrees without shiftfs
This patch removes the workaround due to missing access rights
to the conatiner's mount subtree in case shiftfs is not supported
by the kernel. We achieve that by just bind mounting the whole
subtree in the c_user submodule, reusing the already implemented
shifting functions for shiftfs. Further, the mounting of
proc and sys is fixed in c_vol submodule allowing pivot_root
to work in case user namspaces are active too.