daemon/c_time: Introduce new submodule c_time for time namespace
This patch implements the new time namespace available since
Linux 5.6. It resest the available clocks for boottime and monotonic
by setting the corresponding offsets to the negated current system
values. Thus, for instance uptime will show the time since a container
was started and not the overall system uptime.
Time namspace could not be activated by clone directly but only
by a call to unshare. After unshare the calling process is not
directly part of the new time namspace, to be allowed to set the new
clock offsets. All children will be placed in the new time namspace.
To put the later init process of a container also to the new timespace,
a call to setns using /proc/self/time_for_children does the trick.