daemon/c_cgroups: use unlimited ram as default
Previously the ram limit of a container was set to 1024M.
Since now the ram limit is applied again (see previous patch), we
retain the default case of unlimited ram, by skipping the write to
the memory cgroup in case ram_limit ist set to 0 in the container
config. Further we define 0 as default value in container.proto.