daemon/c_cgroups: properly cleanup cgroups on container stop
The main subsys directories of container cgroups where not removed
early. Thus, for instance the ram limit would be retained during stop
start cycle of a container even if the config was changed.
This was due to missing rmdir of main cgroup subsystem dir after
the recursive cleanup of each subsystem dir. This is fixed now.