daemon/c_user: properly handle dev and cgroups dirs during shift
In case of shiftfs is not supported, function c_user_shift_ids()
chowned files and did bind mounts by jumping to 'shift'.
However, bind mounting is here not appropriate since its a cgroup
subfolder. The code block of chowning files in case of dev and cgroups
before the block which handles all other dirs in case of shiftfs is
not supported. Thus, we assure that we jump to 'success' in both
cases now, if shiftfs is supported or not.