daemon/c_vol: Fixed populating dev by omitting mount points
The content of dev was is copied from host to container. This was
done by recursively copying all the content, thus also sub mounted
file systems like pts where copied to. However, pts contains handles
for ptys which should not be copied to the container. This also
causes errors if an ssh user session to the CML is active during
container start (e.g. in hosted mode).

This is fixed now by reusing the already existing filter for cgroup
devices to also filter out mount points during dev population.