daemon: Allow setup for containers without mount_setup in config
Previously containers without any mount_setup entries did not work, since
there where no rootfs and no init in the target container if setup mode
is activated.

We fixed this by generating an empty (tmpfs) as rootfs for the in
'guestos_config.c' and reuse the allready mapped cservice as fallback
init in 'container.c'.