daemon/c_vol: Fixed link of overlayfs lowerdir during image mount
If the link of lowerdir has to be updated, e.g. for running in setup
mode, the link of an previous mount has to be removed first.
This is due to symlink() does not update an allready existing link.
Therfore, now unlink is called before the actual symlink to lowerdir.