daemon/c_cap: Introduce function to exec with cap_sys_time
To provide the ability to set system time in a privileged container
with userns enabled, we introduce the c_cap_exec_capsystime() function.
It could be used to execute a binary inside the target container
with cap_sys_time in root userns. To keep isolation impact as low as
possible, we switch to the user witch is mapped to the root user in
the userns of the container. Further, we drop all other capabilities,
except CAP_SYS_TIME.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>