Handle token attachment and detachment
If the USB token assigned to a container is attached and/or detached during runtime,
it is necessary to reflect this fact in the corresponding data structures of the scd and cmld.
Therefore, this commit handles the uevents triggered by token de- and attachments
and updates the token state in the scd and cmld.