daemon/uevent: Fix udev uevent handling in case of systemd-udev
The trusted uid of the udev ideamon was assumed to be the pid of the first
process called "udevd" which is correct for eudev. However, in systemd-based
distributions the udevd process is called "systemd-udevd".

We now search in /proc for both binaries and use the one with the smaller
pid, which is likely the udevd started by init in the CML.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>