daemon/time: Introduce new module time for CML maintained time
Since we allow to run processes with CAP_SYS_TIME inside of
containers, we may not trust the real-time clock in CML anymore.
However, this approach is necessary to be able to run an NTP server and
client inside of a container.
However, if the container is malicious we cannot trust the real time clock
anymore. To provide a real-time clock independent time stamp, we rely on the
boot time clock which is time namespaced by the kernel. Further, if out
of sync of real-time clock and boot time based internal clock is detected,
we update the internal btime by a simple internal ntp call.

For certificate validity checks, we can now use the coarse time stamp
provided by time_cml() which is based on the above described implementation.

Further, patches can use this in combination with seccomp notify to emulate
the time() syscall, e.g., used by scd.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>