daemon/cmld: Fixed use after free in cmld_netif_phys_remove_by_name()
A list data pointer was already freed and due to use of mem_free macro set
to NULL at calling list_remove(). However, list_remove expects a data
pointer to be looked up for removing. Correct use here is not to use
list_remove which operates on data pointer but directly use list_unlink()
which just removes the list element.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>