daemon/uevent: register net device by mac address for containers
Maintain a list of mappings between containers and assigned network
interfaces by its mac. This allows to move hot-plugged interfaces to a
container. uevent.h exports a register function to make use of this
from other places. Internally the interface renaming scheme is refactored
to reduce code duplicity for wifi and eth interfaces.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>