daemon/container, c_net: allow to use mac string for netif assignments
The container config file allows to set a repeated string for network
interface names which are assigned/moved to the container net namespace.
This field now accepts a string representation of mac addresses in form
of "00:11:22:33:44:55", too. The protobuf inteface is not touched and the
descission if the string is a mac or ifname is made in c_net submodule.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>