daemon/uevent: replace ifname in cmld's available netifs while renaming
cmld tracks a list of interface names which are allowed to be moved
to c0 (which are not occupied by anoher container). To make sure
that no interface which should be mapped to a container, also the
interface in the traking list of cmld.c needs to updated if an network
interface is renamed. We just check and remove the old interface and
add the renamed in that list.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>