daemon/uevent: use uevent->devtype as prefix for interface renaming
The decision which prefix for renamed interfaces was made by
asking network_is_wifi(). This function depends on sysfs and that
the phy is populated already. However, in case of the uevent
where renaming takes place this is not yet the case. That is why
later also a timed callback for moving the devices is settled.

To avoid the wrong prefix, we use the uevent->devtype string for
renaming now.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>