daemon/uevent: Fixed handling of new physical network interfaces
Due to lately introduced coldboot event filtering, the forwarding of new
physical interfaces to c0 was missed. We reorder the check for network
interfaces after the check if an event was an synthetic coldboot event to
assure that the interface is actually attached from a new device. Thus,
it also can safely added to the cmld tracking list of available devices.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>