daemon/c_net: Do not keep helper child running in netns of c0
Since, veth endpoints of containers are configure in c0, a child of cmld
was created and kept idling in the netns of c0 until the corresponding
container was stopped. This confused logging inside those children, since
cmld does a log rotate in its main.c which is not propagated to that
child. Thus, we simple exit the child when configuration is done and
create a new child for cleanup when a container stops.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>