daemon/c_vol: enabled 'metacopy=on' for overlayfs mounts
For mount_type 'SHARED_RW' an overlayfs is being created with a
temporary 'tmpfs' as writable overlay. Thus, if uid mapping is not
handled by shiftfs but chowning the files, this would cause a copy up
of all files to the ram-based tmpfs. This could lead to exhausting
the memory if the corresponding image is large.
To avoid this, we now use the 'metacopy=on' mount option for
overlayfs. This option just copies the metadata of a file if the real
data is not touched.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>