daemon: Introduce container config signature validation
The API in 'container.c' for container_new() is expanded to also
take buffers for signature and certificate for a container config.
These are given to container_config_new() in 'contrainer_config.c'.
The external API in 'container.c' allows to leave these buffers empty,
and the container_config_new() will deal with reading and storing
corresponding date from/to disk.

The signature verification is also made in container_config_new().
Signature failure result in an empty container object and thus makes
the implementation almost transparent to cmld.c and control.c.

The new feature can be enabled in device.conf 'signed_configs: true'.
Currently, default is set to false to maintain previous behaviour.

Getting signature and certificate through control interface is not
yet implemented and needs to be done in followup commits.
Therefore, container update or creation through control interface do
not work yet if 'signed_configs' is enabled in device.conf.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>