common/cryptfs: use internal ioctl wrapper for non glibc builds
Non glibc std libraries such as musl provide ioctl (int, int, ...)
wrapper. However, dm integrity requests are 'unsigned long int' and
would overflow on a cast to int. Thus, we directly provide an syscall
wrapper in the cryptfs.c instead of using the ioctl wrapper of the
std library in that case.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>