tpm2d: allow to start tpm2d without initializing keys for attestation
In case of disk encryption, tpm2d needs to be started early were the
data partition which is used to store key blobs is not yet available.
Hence, we need to start tpm2d twice in this scenario.
The first time to initialize disk encryption and the second time when
encrypted data partition is available. Then keys could be stored on
the encrypted partition.
This patch allows to start tpm2d with cmdline parameter -n (--nokeys)
to provide initial mode for using to unlock the disk encryption key.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>