service/service: Retrieve audit messages
This commit modifies the cservice such that it
either forks a message handling process
if launched with PID == 1 and no init is given
or directly enters an event loop such that it can
be used in conjunction with init systems like systemd
or busybox init.

Currently, the message handler only processes audit record
related messages. This can be extended in future commits.

Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>