daemon/c_cap: exec with cap_systime of ntpd using uid wrapper
Use LD_PRELOAD magic provided by libuid_wrapper.so to trick ntpd into
thinking of running as root. This is needed since, ntpd exits if uid
check is not 0. Since we switch to the root user of the user
namespace, but not joining the user namespace, this mapping is not
provided by the userns.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>