daemon/cc_mode: Reduce CML interface to be minimal
In order to prepare trust|me to be evaluated according to CC
we need to reduce all interfaces to the absolute minimum.

This commit reduces the interfaces that are implemented via the
control and service sockets. To avoid having to maintain a hard
fork, a new preprocesser flag, was already introduced.
This flag will effectively filter commands that the cmld services.

In order for clients to be able to implement the reduced interfaces
we supply the specification in form of adapted .proto files.

Signed-off-by: Christian Epple <christian.epple@aisec.fraunhofer.de>
Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>