daemon/container_config: added option for mac filter on phys netifs
We provide a new config message in container.proto for physical network
interface assignments. Since mac filter needs to put on those interfaces,
a repeated string alone is not sufficient anymore. The corresponding
netif is now part of the new ContainerPnetConfig message. Further, a
repeated string for allowed client mac addresses is included.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>