common/cryptfs: make integrity setup optionally by parameter
We use 'cryptfs_setup_volume_new()' also in tpm2d for FDE. There we
depend on the ability to do block-wise on the fly encryption and
therefore need to maintain the previous behaviour of a dm-crypt only
device. To achieve that, we introduce an additional parameter to just
activate integrity explicitly in 'cryptfs_setup_volume_new()'.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>