common/cryptfs: Propper dm_crypt device setup on integrity device
This is the real for integrity device setup and also inverts the
Quick-Fix (commit b7238cd5e889fca73eb2144e55472c7117754aa3).

We now use a separate meta device for integrity data. Thus, we can
use the blk size of the real block device which should be encrypted.

We read out the data size of the super block of the integrity device
to check if we need initial setup of meta data.
Further, properly initialize all data of the dm-crypt device by writing
block wise zeros to the dm-crypt device.
Otherwise not block aligned writes will cause I/O errors, since no
integrity data exist yet. If the write is not filling the whole block,
a read is performed before write and thus will result in the I/O error.

Further, some usage of closed file descriptors are fixed.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>