daemon/audit,c_audit: restructured and fixed protobuf memory handling
mem_free of protobuf structures lead to a memory corruption in the
daemon/audit submodule. We fixed this by resturcteing and using the
corresponding protobuf_free methods to free the protobuf structures.
Further, some frees of buffers allocated by getline and other
memory related cleanups are provided in this commit.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>