daemon/container, c_audit: restructured start to allow early handlers
In some cases, work needs to be done in a child before actually cloned
to new namespaces. Especially, in the audit use case loginuids can not
be written from user namespace with a uid outside of the namespace.
Thus, for session tracing we need an early handler (which is also part
of this commit) to do this.
The actual clone to the new container child is then made by that early
child. So to say we do a double fork.

This could also be used to provide further start_child_early() handlers
for doing work in the CML root namespace without blocking the main
cmld process which is needed early during the container startup.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>