daemon/container, c_vol: fixed cleanup and rebooting of c_vol
Instead of keeping the state of c_vol during reboots we now just keep
the dm crypt and integrity devices in the cleanup of c_vol.
To be able to reuse the allready setup dm devices we have to set
a dummy key for the restart process inside the container struct.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>