daemon/tss: use TPM2D_BINARY_NAME define for lsb builds
In tpm2d's lsb Makefile the tpm2d binary is installed to cml-tpm2d. This
resulted in a fatal abort. Thus, we make use of a define for the binary
name now. This define is set to cml-tpm2d in Makefile_lsb so that the cmld
can find the corresponding {cml-}tpm2d binary.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>