daemon: Rename token pairing state flag file
Flag file was named "is_init" which created ambigiouty.

Now, initialization of a token refers only to its state within the
SCD, i.e. whether it has been enumerated and initialized.
The flag file is now named "token_is_paired". Pairing refers to the
provisioning state of a token, i.e. whether it has been provisioned
with a platform-bound authentication code which consists of a user pin
and a pairing secret.

Signed-off-by: Christian Epple <christian.epple@aisec.fraunhofer.de>