c_net: Fix veth cleanup in c0
Currently, the veth cleanup code for c0 joins the net namespace only.
As the cleanup code determines existense of an interface based on /sys/class/net
it is necessary to also join the mount namespace of c0
s.t. the sysfs reflects the interfaces present in the net namespace.

Also, the cleanup routine is currently called twice,
this commit removes the additional loop.

Signed-off-by: Felix Wruck <felix.wruck@aisec.fraunhofer.de>