daemon/c_net: provide function to move/bridge physical netifs to c_net
If the physical interface needs a mac filter to be applied a new
veth pair is generated and bridged to the physical interface inside
the root netns of the CML. The corresponding veth endpoint is moved
to the container. Thus, it is possible to apply the mac filter on
the bridge inside the root netns totally transparent to the container.
If no mac filter is configured, the new function just moves the
physical network interface directly to the container's netns.

We designed the new functiona in a way to be used externaly from
container.c and uevent.c when we switch to container_pnet_cfg_t
instead of ifnames there, too. This however is left to furhter commits.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>